% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{all_are_distinct}
\alias{all_are_distinct}
\title{Test if all elements of a vector are unique}
\usage{
all_are_distinct(x, empty_value = FALSE)
}
\arguments{
\item{x}{any object that can be handled by \code{\link[=unique]{unique()}} (usually a vector or
list)}

\item{empty_value}{Value to return if function is called on a vector of
length 0 (e.g. \code{NULL}, \code{numeric()}, ...)}
}
\value{
\code{TRUE} or \code{FALSE}
}
\description{
Test if all elements of a vector are unique
}
\examples{
all_are_distinct(c(1,2,3))
all_are_distinct(c(1,1,1))
}
