% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{as_scalar}
\alias{as_scalar}
\title{Convert vector if identical elements to scalar}
\usage{
as_scalar(x)
}
\arguments{
\item{x}{any object that can be handled by \code{\link[=unique]{unique()}} (usually a vector or
list)}
}
\value{
A scalar of the same type as \code{x}
}
\description{
Returns \code{unique(x)} if all elements of \code{x} are identical, throws an error if
not.
}
