% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{assert}
\alias{assert}
\title{Assert a condition}
\usage{
assert(cond, ..., call. = FALSE, domain = NULL)
}
\arguments{
\item{cond}{\code{TRUE} or \code{FALSE} (without any attributes). \code{FALSE} will throw
an exception with an automatically constructed error message (if \code{...}
was not supplied). Anything else will throw an exception stating that
\code{cond} was not valid.}

\item{...}{passed on to \code{\link[=stop]{stop()}}}

\item{call.}{passed on to \code{\link[=stop]{stop()}}}

\item{domain}{passed on to \code{\link[=stop]{stop()}}}
}
\value{
TRUE on success
}
\description{
A simpler and more efficient for \code{\link[base:stopifnot]{base::stopifnot()}} that has an easy
mechanism for supplying custom error messages. As opposed to \code{stopifnot()},
\code{assert()} only works with a single (scalar) assertions.
}
\examples{

\dontrun{
assert(1 == 1)
assert(1 == 2)
}


}
