% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{assert_all}
\alias{assert_all}
\title{Assert all elements of a vector or list are \code{TRUE}}
\usage{
assert_all(..., call = sys.call(-1))
}
\arguments{
\item{...}{\code{logical} vectors or \code{lists} with only \code{logical} elements}
}
\value{
\code{TRUE}
}
\description{
Check if all elements of \code{x} are \code{TRUE}, and throw an informative warning that
contains the position and name (if any) of the element
}
\examples{
try(
  assert_all(c("this is true" = TRUE, "this is FALSE" = FALSE, FALSE))
  assert_all(
    list(a = c(TRUE, FALSE), b = TRUE),
    list(TRUE)
  )
)
}
\seealso{
\code{\link[=assert]{assert()}}
}
