% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{camelCase}
\alias{camelCase}
\title{Turn a character vector to camelCase}
\usage{
camelCase(x, sep_pattern = "_|\\\\s")
}
\arguments{
\item{x}{a \code{character} vector}

\item{sep_pattern}{a \code{regex} pattern to match separators}
}
\value{
a \code{character} vector that follows camelCase guidelines
}
\description{
\strong{EXPERIMENTAL}
}
\examples{
camelCase("foo_bar")
}
