% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{comma}
\alias{comma}
\title{Collapse text vectors with a comma}
\usage{
comma(..., collapse = ", ")
}
\arguments{
\item{x}{\code{character} vector}
}
\value{
a \code{character} scalar
}
\description{
Collapse text vectors with a comma
}
