% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{commaset}
\alias{commaset}
\title{Collapse text vectors with a comma (no duplicates)}
\usage{
commaset(..., collapse = ", ")
}
\arguments{
\item{x}{\code{character} vector}
}
\value{
a \code{character} scalar
}
\description{
Collapse text vectors with a comma (no duplicates)
}
