% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{equalish}
\alias{equalish}
\alias{equalish_frac}
\title{Check for equality within a tolerance level}
\usage{
equalish(x, y, tolerance = .Machine$double.eps^0.5)

equalish_frac(x, y, tolerance = .Machine$double.eps^0.5)
}
\arguments{
\item{x, y}{\code{numeric} vectors}

\item{tolerance}{\code{numeric} scalar. tolerance level (absolute value). Defaults
to \code{.Machine$double.eps^0.5} which is a sensible default for comparing
floating point numbers.}
}
\value{
\code{equalish()} returns TRUE if the absolute difference between \code{x} and
\code{y} is less than \code{tolerance}.

\code{equalish_frac()} returns \code{TRUE} if the relative difference between
\code{x} and \code{y} is smaller than \code{tolerance}. The relative difference is
defined as \code{abs(x - y) / pmax(abs(x), abs(y))}. If both \code{x} and \code{y} are
\code{0} the relative difference is not defined, but this function will still
return \code{TRUE}.
}
\description{
Check for equality within a tolerance level
}
\examples{
a <- 0.7
b <- 0.2
a - b == 0.5
equalish(a - b, 0.5)
equalish_frac(1000, 1010, tolerance = 0.01)
equalish_frac(1000, 1010, tolerance = 0.009)
equalish_frac(0, 0)
}
\seealso{
\link{.Machine}
}
