% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{fmt_class}
\alias{fmt_class}
\alias{class_fmt}
\title{Format class of an object}
\usage{
fmt_class(x, open = "<", close = ">")

class_fmt(x, ignore = NULL, open = "<", close = ">")
}
\arguments{
\item{x}{any \R object}

\item{open, close}{opening and closing bracket to use for formatting}

\item{ignore}{subclasses to ignore}
}
\description{
Format class of an object
}
\examples{
x <- iris
class(x) <- c("flowers", "data.frame")
class_fmt(x)
fmt_class(class(x))
}
