% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{is_blank}
\alias{is_blank}
\title{Check if a String is Blank}
\usage{
is_blank(x)
}
\arguments{
\item{x}{Any \R Object.}
}
\value{
either \code{TRUE} or \code{FALSE}
}
\description{
Check wheter a character vector contains only of spaces
}
