% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{is_empty}
\alias{is_empty}
\title{Check if Object has length 0}
\usage{
is_empty(x)
}
\arguments{
\item{x}{Any \R Object.}
}
\value{
either \code{TRUE} or \code{FALSE}
}
\description{
Check wheter an object is either \code{TRUE} or \code{FALSE}.
}
