% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{is_equal_length}
\alias{is_equal_length}
\title{Check if Objects have the same length}
\usage{
is_equal_length(...)
}
\arguments{
\item{...}{Any number of \R Objects.}
}
\value{
either \code{TRUE} or \code{FALSE}
}
\description{
Check if Objects have the same length
}
