% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{is_error}
\alias{is_error}
\alias{is_try_error}
\alias{is_scalar}
\alias{is_POSIXct}
\alias{is_scalar_POSIXct}
\alias{is_POSIXlt}
\alias{is_scalar_POSIXlt}
\alias{is_POSIXt}
\alias{is_scalar_POSIXt}
\alias{is_Date}
\alias{is_scalar_Date}
\alias{is_scalar_list}
\alias{is_scalar_atomic}
\alias{is_scalar_logical}
\alias{is_scalar_integer}
\alias{is_scalar_factor}
\alias{is_scalar_numeric}
\alias{is_scalar_character}
\alias{is_vector}
\alias{is_bool}
\alias{is_scalar_integerish}
\alias{is_n}
\alias{is_n0}
\alias{is_not_distinct_from}
\alias{is_distinct_from}
\alias{is_windows_path}
\alias{is_dir}
\alias{is_empty_dir}
\title{Convenient predicate functions}
\usage{
is_error(x)

is_try_error(x)

is_scalar(x)

is_POSIXct(x)

is_scalar_POSIXct(x)

is_POSIXlt(x)

is_scalar_POSIXlt(x)

is_POSIXt(x)

is_scalar_POSIXt(x)

is_Date(x)

is_scalar_Date(x)

is_scalar_list(x)

is_scalar_atomic(x)

is_scalar_logical(x)

is_scalar_integer(x)

is_scalar_factor(x)

is_scalar_list(x)

is_scalar_numeric(x)

is_scalar_character(x)

is_vector(x)

is_bool(x)

is_scalar_integerish(x)

is_n(x)

is_n0(x)

is_not_distinct_from(x, y)

is_distinct_from(x, y)

is_windows_path(x)

is_dir(x)

is_empty_dir(x)
}
\arguments{
\item{x}{Any \R Object.}
}
\value{
either \code{TRUE} or \code{FALSE}
}
\description{
Convenient predicate functions
}
\seealso{
https://modern-sql.com/feature/is-distinct-from
}
