% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{is_integerish}
\alias{is_integerish}
\title{Check if Object is Integer-like}
\usage{
is_integerish(x, na_rm = FALSE)
}
\arguments{
\item{x}{Any \R Object.}
}
\value{
either \code{TRUE} or \code{FALSE}
}
\description{
Check wheter an object is either \code{TRUE} or \code{FALSE}.
}
