% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{is_scalar_bool}
\alias{is_scalar_bool}
\title{Check if Object is a Boolean}
\usage{
is_scalar_bool(x)
}
\arguments{
\item{x}{Any \R Object.}
}
\value{
either \code{TRUE} or \code{FALSE}
}
\description{
Check wheter an object is either \code{TRUE} or \code{FALSE}.
}
