% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{path_tidy}
\alias{path_tidy}
\title{Clean up paths to make them comparable, inspired by \code{\link[fs:path_tidy]{fs::path_tidy()}}}
\usage{
path_tidy(x)
}
\arguments{
\item{x}{\code{character} vector}
}
\value{
a \code{character} vector
}
\description{
Clean up paths to make them comparable, inspired by \code{\link[fs:path_tidy]{fs::path_tidy()}}
}
