% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{rd_describe_str}
\alias{rd_describe_str}
\title{Create description string for listlike R objects Rd documentation}
\usage{
rd_describe_str(x)
}
\arguments{
\item{x}{any \R object}
}
\description{
\strong{EXPERIMENTAL}
}
