% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfmisc-package.R
\docType{package}
\name{sfmisc-package}
\alias{sfmisc}
\alias{sfmisc-package}
\title{Private utility functions}
\description{
Commonly used utility functions without dependencies that can easily be embedded in R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://s-fleck.github.io/sfmisc/}
}

}
\author{
\strong{Maintainer}: Stefan Fleck \email{stefan.b.fleck@gmail.com} (\href{https://orcid.org/0000-0003-3344-9851}{ORCID})

}
