% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{validate}
\alias{validate}
\title{Simple data validation}
\usage{
validate(...)
}
\arguments{
\item{...}{Arbitrary expressions that evaluate to either \code{TRUE} or \code{FALSE}.
Expressions that evaluate to anything else, will be counted as \code{FALSE} and
throw a warning. You can name the expressions to generate nice labels
(but that's usually not necessary).}
}
\value{
a named \code{logical} vector that is guranteed to have no \code{NAs}.
}
\description{
\strong{EXPERIMENTAL} Minimalistic mutli-criteria data validation. Designed to
work nicely together with \code{assert_all()} (see examples)
}
\examples{
validation <- validate(
  all(iris$Petal.Length < 5),
  `all species are valid` = all(iris$Species \%in\% c("setosa", "versicolor", "virginica")),
  `all sepals are small` = all(iris$Sepal.Length < 1)
)

# validate can be used `with()` for added convenience:
validation <- with(iris, validate(
  all(Petal.Length < 5),
  `all species are valid` = all(Species \%in\% c("setosa", "versicolor", "virginica")),
  `all sepals are small` = all(Sepal.Length < 1)
))

# validate works together with assert_all to produce nice errors
try(
  assert_all(validation)
)
}
