% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sfmisc.R
\name{walk}
\alias{walk}
\title{Apply a function to each element of a list or vector (without returning anything)}
\usage{
walk(.x, .f, ...)
}
\arguments{
\item{.x}{an atomic \code{vector} or a \code{list}}

\item{.f}{a \code{function}}

\item{...}{arguments passed on to \code{.f}}
}
\value{
\code{.x}
}
\description{
Like lapply, but return \code{.x} instead of the results of \code{.f}
}
\examples{
walk(month.name, print)
}
